+++ src/meta/assessment/originalDataPoint/odps/test/mock/commonProps.ts
@@ -1,0 +1,10 @@
+export const commonProps = {
+  id: 1400,
+  countryIso: 'X01',
+  year: 2011,
+  dataSourceAdditionalComments: 'Add comments here',
+  dataSourceMethods: ['nationalForestInventory'],
+  dataSourceReferences: '',
+  description: '',
+  idLegacy: null as unknown,
+}
