--- src/meta/assessment/originalDataPoint/odps/test/mockODP.ts
@@ -1,661 +1,0 @@
-import { OriginalDataPoint } from 'meta/assessment/originalDataPoint/originalDataPoint'
-
-const commonProps = {
-  id: 1400,
-  countryIso: 'X01',
-  year: 2011,
-  dataSourceAdditionalComments: 'Add comments here',
-  dataSourceMethods: ['nationalForestInventory'],
-  dataSourceReferences: '',
-  description: '',
-  idLegacy: null as unknown,
-}
-
-export const odp1 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: '6c611c99-9eab-4369-a56b-16c671e491da',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp2 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '50.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: '6c611c99-9eab-4369-a56b-16c671e491da',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp3 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '50.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: '6c611c99-9eab-4369-a56b-16c671e491da',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp4 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: null,
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '50.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: null,
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: null,
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: null,
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'd747092c-bf76-42c0-afd9-dced0e5d8a75',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp5 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '5.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '50.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: null,
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: null,
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: null,
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: '589e3b51-fbef-4998-970f-a2b3133493e2',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp6 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '50.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '100.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: null,
-      otherWoodedLandPercent: '50.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'b27d8f15-6a69-4b6b-805c-c2dfce85ddae',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp7 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '5',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: null,
-      otherWoodedLandPercent: '50.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'b27d8f15-6a69-4b6b-805c-c2dfce85ddae',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp8 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'e4470f48-11f8-460f-b400-ba4b939d8a6e',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp9 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: '5',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: '100.00',
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: '0',
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: '0.00',
-      otherWoodedLandPercent: '100.00',
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'e4470f48-11f8-460f-b400-ba4b939d8a6e',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp10 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: null,
-      forestPlantationPercent: '5.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: null,
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: null,
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: null,
-      otherWoodedLandPercent: null,
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: 'f3074aa1-c5b4-4984-86e9-87eb97648619',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
-
-export const odp11 = {
-  ...commonProps,
-  nationalClasses: [
-    {
-      area: '1500.00',
-      name: 'Open forest',
-      uuid: '02b381ca-2c39-49f7-9fef-80a3463f8f6b',
-      definition: 'Add definition here',
-      forestPercent: '5.00',
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: '5.00',
-      forestPlantationPercent: '5.00',
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '50.00',
-    },
-    {
-      area: '900.00',
-      name: 'Closed forest',
-      uuid: '3986836e-eac8-4287-a625-11df18a60d01',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '50.00',
-      otherWoodedLandPercent: null,
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: '5.00',
-      forestNaturalForestOfWhichPrimaryForestPercent: null,
-    },
-    {
-      area: '550.00',
-      name: 'Plantations',
-      uuid: '42b9b108-c655-40f3-a865-934346dfa859',
-      definition: '',
-      forestPercent: null,
-      forestNaturalPercent: '0.00',
-      otherWoodedLandPercent: null,
-      forestPlantationPercent: null,
-      otherPlantedForestPercent: '0.00',
-      forestPlantationIntroducedPercent: null,
-      forestNaturalForestOfWhichPrimaryForestPercent: '0',
-    },
-    {
-      area: '300.00',
-      name: 'Woodland',
-      uuid: '578bb1d6-961a-48c1-b489-877641f3c17f',
-      definition: '',
-      forestPercent: null,
-      otherWoodedLandPercent: null,
-    },
-    {
-      name: '',
-      definition: '',
-      uuid: '012d5db6-3990-43f0-bc12-c99fe8467bbb',
-      placeHolder: true,
-    },
-  ],
-} as OriginalDataPoint
